Add tests for Controller
